add spinlock example combined with timers
A new data-protected block has been added to the map element. It is
essential that the bpf_spin_lock data is handled with care, as the
verification success is dependent upon the padding of the element map
structure.
In an ebpf kernel program, a spinlock-protected block is prohibited from
including any helper functions or kfunc calls.

On the user space side, the BPF_F_LOCK flag for map operations should be
utilized to enforce atomicity.

Signed-off-by: Andrea Mayer <andrea.mayer@uniroma2.it>