WIP implement packet rate measurement based on slotted window approach
Packet rate measurement is accomplished using the slotted window
approach. Currently, we have only one slotted window that evaluates the
average packet rate for all flows generated by the node. This is
obviously an example: in a real use case scenario, the average packet
rate is evaluated on a per-flow basis.

Therefore, in this commit we do not address packet/flow classification
in any way, it is still left as TODO.

Signed-off-by: Andrea Mayer <andrea.mayer@uniroma2.it>