add a simple xdp program to test fwd performance
a simple xdp program is added to test the performance in a network
topology made of two server connected back-to-back (ring configuration)
as follows:

    +------------------+      +------------------+
    |        r0        |      |        r1        |
    |                  |      |                  |
    |       vrf0       |      | xdp_redirect     |
 fd00::1/64   *veth0 +----------+ veth0    fd00::2/64
    |        xdp_pass  |      |                  |
    |        xdp_pass  |      |                  |
 fd01::1/64   *veth1 +----------+ veth1    fd01::2/64
    |       vrf1       |      | xdp_redirect     |
    |                  |      |                  |
    +------------------+      +------------------+

The XDP redirect program is configured on both veth0 and veth1 of node
r1. Packets coming from veth0 in r0 are received from veth0 in r1 and
sent back to r0 using veth1.
The reverse path is similar to the direct one.

Signed-off-by: Andrea Mayer <andrea.mayer@uniroma2.it>