update podman README.md file
To guarantee correct ownership and prevent permission errors, it's
important to execute all steps, beginning with cloning the repository,
as the root user. This involves switching to root using sudo su prior to
cloning, so that the repository will be owned by root and associated
with the root group. Working as root for the entire process streamlines
deployment and sidesteps ownership issues, but it demands careful
handling because of the complete control it provides.

Note: Working with root privileges can potentially damage the system if
commands are executed incorrectly. Be cautious and ensure you understand
each step to avoid unintentional system modifications or security risks.

Signed-off-by: Andrea Mayer <andrea.mayer@uniroma2.it>