Remove sudo from setup scripts to run as root for simplified env setup
Avoid using sudo in the script since many Linux distributions do not
support it by default.

To ensure the scripts work correctly, it is preferable to run it
directly as the root user, especially since security is not a concern in
our controlled environment.

While using root is generally discouraged and not recommended for
production systems, in this case, it simplifies setup and allows us to
install numerous packages and configure privileged containers that can
access host devices like /dev/kvm without worrying about device
permissions, ownership, or access control.

Our goal is to create a convenient environment for testing and
development within a laboratory setting, where the usual security
considerations are relaxed.
However, it's important to remember that running as root is discouraged
in most real-world scenarios due to security risks.

Signed-off-by: Andrea Mayer <andrea.mayer@uniroma2.it>